+++ src/@shared/constants/lotto.js
