+++ src/components/Lotto/style.js
@@ -1,0 +1,11 @@
+import styled from "@emotion/styled";
+
+export const List = styled.li`
+  display: flex;
+  align-items: center;
+`;
+
+export const TicketIcon = styled.span`
+  margin-right: 10px;
+  font-size: 24px;
+`;
