+++ src/components/LottoContainer/style.js
@@ -1,0 +1,11 @@
+import styled from "@emotion/styled";
+
+// eslint-disable-next-line import/prefer-default-export
+export const Header = styled.h2`
+  display: flex;
+  justify-content: space-between;
+  margin-top: 20px;
+  font-size: 16px;
+  font-weight: normal;
+  align-items: center;
+`;
