+++ src/components/Main/style.js
@@ -1,0 +1,7 @@
+import styled from "@emotion/styled";
+
+// eslint-disable-next-line import/prefer-default-export
+export const MainContainer = styled.main`
+  width: 23vw;
+  min-width: 400px;
+`;
