--- src/contexts/LottoContext.js
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const LottoContext = createContext([]);
-
-export default LottoContext;
